+++ Robust.Client/UserInterface/RichText/MarkupTagManager.cs
@@ -16,7 +16,7 @@ public sealed class MarkupTagManager
     /// <summary>
     /// Tags defined in engine need to be instantiated here because of sandboxing
     /// </summary>
-    private readonly Dictionary<string, IMarkupTag> _markupTagTypes = new IMarkupTag[] {
+    private readonly Dictionary<string, IMarkupTagHandler> _markupTagTypes = new IMarkupTagHandler[] {
         new BoldItalicTag(),
         new BoldTag(),
         new BulletTag(),
@@ -44,13 +44,13 @@ public sealed class MarkupTagManager
 
     public void Initialize()
     {
-        foreach (var type in _reflectionManager.GetAllChildren<IMarkupTag>())
+        foreach (var type in _reflectionManager.GetAllChildren<IMarkupTagHandler>())
         {
             //Prevent tags defined inside engine from being instantiated
             if (_engineTypes.Contains(type))
                 continue;
 
-            var instance = (IMarkupTag)_sandboxHelper.CreateInstance(type);
+            var instance = (IMarkupTagHandler)_sandboxHelper.CreateInstance(type);
             _markupTagTypes[instance.Name.ToLower()] = instance;
         }
 
@@ -60,22 +60,48 @@ public sealed class MarkupTagManager
         }
     }
 
+    [Obsolete("Use GetMarkupTagHandler")]
     public IMarkupTag? GetMarkupTag(string name)
+    {
+        return _markupTagTypes.GetValueOrDefault(name) as IMarkupTag;
+    }
+
+    public IMarkupTagHandler? GetMarkupTagHandler(string name)
     {
         return _markupTagTypes.GetValueOrDefault(name);
     }
 
-    public bool TryGetMarkupTag(string name, Type[]? tagsAllowed, [NotNullWhen(true)] out IMarkupTag? tag)
+    /// <summary>
+    /// Attempt to get the tag handler with the corresponding name.
+    /// </summary>
+    /// <param name="name">The name of the tag, as specified by <see cref="IMarkupTag.Name"/></param>
+    /// <param name="tagsAllowed">List of allowed tag types. If null, all types are allowed.</param>
+    /// <param name="handler">The instance responsible for handling tags of this type.</param>
+    /// <returns></returns>
+    public bool TryGetMarkupTagHandler(string name, Type[]? tagsAllowed, [NotNullWhen(true)] out IMarkupTagHandler? handler)
     {
         if (_markupTagTypes.TryGetValue(name, out var markupTag)
             // Using a whitelist prevents new tags from sneaking in.
             && (tagsAllowed == null || Array.IndexOf(tagsAllowed, markupTag.GetType()) != -1))
         {
-            tag = markupTag;
+            handler = markupTag;
             return true;
         }
 
-        tag = null;
+        handler = null;
         return false;
+    }
+
+    [Obsolete("Use TryGetMarkupTagHandler")]
+    public bool TryGetMarkupTag(string name, Type[]? tagsAllowed, [NotNullWhen(true)] out IMarkupTag? tag)
+    {
+        if (!TryGetMarkupTagHandler(name, tagsAllowed, out var handler) || handler is not IMarkupTag cast)
+        {
+            tag = null;
+            return false;
+        }
+
+        tag = cast;
+        return true;
     }
 }